Fix the nav on the screen
I had to get rid of the CSS grid because it turns out it's much easier
to implement the thing using just a plain large margin for the content.

I could probably achieve the same with grid by setting the nav column
to some fixed value instead of using `fr`s.